+++ src/main.jsx
@@ -3,14 +3,15 @@ import ReactDOM from "react-dom/client";
 import App from "./App.jsx";
 import store from "./redux/store";
 import { Provider } from "react-redux";
-import { ToastContainer } from "react-toastify";
+import { BrowserRouter } from "react-router-dom";
 import "react-toastify/dist/ReactToastify.css";
 
 ReactDOM.createRoot(document.getElementById("root")).render(
   <React.StrictMode>
-    <Provider store={store}>
+    <BrowserRouter>
+      <Provider store={store}>
         <App />
-        <ToastContainer />
       </Provider>
+    </BrowserRouter>
   </React.StrictMode>,
 );
